+++ ChangeLog.md
@@ -1,14 +1,20 @@
-# Version 1.6
+# Version 1.7.0
+
+- Updated Mutagen to be compatible with header version 1.71
+- Fixed RemoveEntriesFromLists output, it was failing because it was not importing the load order properly
+- New Program PluginCheckerForMerge
+
+# Version 1.6.0
 
 - New program MainMenuRandomizerAssistant
 - New program ScriptEditor
 
-# Version 1.5
+# Version 1.5.0
 
 - New program DetectProblematicRecordsForMutagen
 - New program DetectFormLinksInRecords
 
-# Version 1.4
+# Version 1.4.0
 
 - New program CombineFolders
 - New program RemoveEntriesFromLists